add swimlane certificate check

+++ swimlane_environment_validator/lib/verify_pip.py
@@ -2,6 +2,7 @@
 import subprocess
 import sys
 import shutil
+import os
 
 import swimlane_environment_validator.lib.config as config
 import swimlane_environment_validator.lib.log_handler as log_handler
@@ -27,7 +28,7 @@ def create_virtual_env():
         logger.error("Something went wrong with trying to create a virtualenv. Is python3 and python3-venv installed?")
         return False
 
-    if config.arguments.pip_config:
+    if config.arguments.pip_config and os.path.isfile(config.arguments.pip_config):
         shutil.copyfile(config.arguments.pip_config, 'pip-install-test-venv/pip.conf')
 
     if sp.returncode != 0:
